refactor(net): split darwin datapath loop
